LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-form-expected.txt


FAIL <form> submission fires navigate event assert_equals: expected (object) Element node <form id="form" method="post" action=""></form> but got (undefined) undefined
